+++ src/onegov.onboarding/onegov/onboarding/models/town_assistant.py
@@ -163,7 +163,7 @@ class TownAssistant(Assistant):
             })
 
             self.app.es_perform_reindex()
-            self.app.send_email(
+            self.app.send_transactional_email(
                 subject=title,
                 receivers=(user, ),
                 content=welcome_mail,